start implementing source command
