+++ crates/shell/src/main.rs
@@ -5,16 +5,16 @@ use std::rc::Rc;
 use anyhow::Context;
 use clap::Parser;
 use deno_task_shell::parser::debug_parse;
-use deno_task_shell::{
-    execute_sequential_list, AsyncCommandBehavior, ExecuteResult, ShellCommand, ShellPipeReader,
-    ShellPipeWriter, ShellState,
-};
+use deno_task_shell::{ShellCommand, ShellState};
 use rustyline::error::ReadlineError;
 use rustyline::{CompletionType, Config, Editor};
 
 mod commands;
 mod completion;
+mod execute;
 mod helper;
+
+pub use execute::execute;
 
 fn commands() -> HashMap<String, Rc<dyn ShellCommand>> {
     HashMap::from([
@@ -30,41 +30,11 @@ fn commands() -> HashMap<String, Rc<dyn ShellCommand>> {
             "unalias".to_string(),
             Rc::new(commands::AliasCommand) as Rc<dyn ShellCommand>,
         ),
+        (
+            "source".to_string(),
+            Rc::new(commands::SourceCommand) as Rc<dyn ShellCommand>,
+        ),
     ])
-}
-
-async fn execute(text: &str, state: &mut ShellState) -> anyhow::Result<i32> {
-    let list = deno_task_shell::parser::parse(text);
-
-    let mut stderr = ShellPipeWriter::stderr();
-    let stdout = ShellPipeWriter::stdout();
-    let stdin = ShellPipeReader::stdin();
-
-    if let Err(e) = list {
-        let _ = stderr.write_line(&format!("Syntax error: {}", e));
-        return Ok(1);
-    }
-
-    // spawn a sequential list and pipe its output to the environment
-    let result = execute_sequential_list(
-        list.unwrap(),
-        state.clone(),
-        stdin,
-        stdout,
-        stderr,
-        AsyncCommandBehavior::Wait,
-    )
-    .await;
-
-    match result {
-        ExecuteResult::Continue(exit_code, changes, _) => {
-            // set CWD to the last command's CWD
-            state.apply_changes(&changes);
-            std::env::set_current_dir(state.cwd()).context("Failed to set CWD")?;
-            Ok(exit_code)
-        }
-        ExecuteResult::Exit(_, _) => Ok(0),
-    }
 }
 
 #[derive(Parser)]
@@ -92,10 +62,6 @@ async fn interactive() -> anyhow::Result<()> {
 
     let helper = helper::ShellPromptHelper::default();
     rl.set_helper(Some(helper));
-
-    // let h = ShellCompleter {};
-
-    // rl.set_helper(Some(h));
 
     let mut state = init_state();
